feat:  /questions/{questionId} API 응답 구조 변경 반영
