--- src/types/answer.ts
@@ -1,17 +1,0 @@
-export interface IAnswer {
-    answerId: number;
-    type: string;
-    account: {
-        name: string;
-        email: string;
-    };
-    content: string;
-    heartCount: number;
-    createdAt: string; // string / Date
-    modifiedAt: string;
-    question: {
-        id: number;
-        category: string;
-        title: string;
-    };
-}
